array: fixing array and passing test for non-uniform string els
